CH-6-Exercise

+++ CH-4-Exercise/16_Question.c
@@ -4,20 +4,20 @@
 
 int main()
 {
-    float num1, num2, num3, average;
+    int num1, num2, num3, average;
 
     printf("Enter the first number: ");
-    scanf("%f", &num1);
+    scanf("%d", &num1);
 
     printf("Enter the second number: ");
-    scanf("%f", &num2);
+    scanf("%d", &num2);
 
     printf("Enter the third number: ");
-    scanf("%f", &num3);
+    scanf("%d", &num3);
 
     average = (num1 + num2 + num3) / 3;
 
-    printf("The average of %f, %f, and %f is %f", num1, num2, num3, average);
+    printf("The average of %d, %d, and %d is %d", num1, num2, num3, average);
 
     return 0;
 }
